+++ src/main/java/com/knots/backend/models/dtos/GeometryDto.java
@@ -3,7 +3,7 @@ package com.knots.backend.models.dtos;
 import java.util.List;
 
 public record GeometryDto (
-        List<List<Integer>> vertex_positions,
-        List<List<Integer>> arrows,
-        List<List<Integer>> crossing_specs
+        List<List<Long>> vertex_positions,
+        List<List<Long>> arrows,
+        List<List<Long>> crossing_specs
 ) {}
