+++ src/main/java/com/knots/backend/repositories/CrossingSpecsRolfRepo.java
@@ -5,6 +5,6 @@ import com.knots.backend.models.entities.CrossingSpecsRolf;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface CrossingSpecsRolfRepo extends JpaRepository<CrossingSpecsRolf, Integer> {
-    List<CrossingSpecsRolf> findAllByDiagramIdOrderByCrossingIdAsc(Integer diagramId);
+public interface CrossingSpecsRolfRepo extends JpaRepository<CrossingSpecsRolf, Long> {
+    List<CrossingSpecsRolf> findAllByDiagramIdOrderByCrossingIdAsc(Long diagramId);
 }
