+++ src/main/java/com/knots/backend/repositories/InvariantRolfRepo.java
@@ -4,6 +4,6 @@ import com.knots.backend.models.entities.InvariantsRolf;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvariantRolfRepo extends JpaRepository<InvariantsRolf, Integer> {
-    InvariantsRolf findByKnotId(Integer knotId);
+public interface InvariantRolfRepo extends JpaRepository<InvariantsRolf, Long> {
+    InvariantsRolf findByKnotId(Long knotId);
 }
