+++ src/main/java/com/knots/backend/repositories/VerticesAndArrowsRolfRepo.java
@@ -5,6 +5,6 @@ import com.knots.backend.models.entities.VerticesAndArrowsRolf;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface VerticesAndArrowsRolfRepo extends JpaRepository<VerticesAndArrowsRolf, Integer> {
-    List<VerticesAndArrowsRolf> findAllByDiagramIdOrderByStartPointAsc(Integer diagramId);
+public interface VerticesAndArrowsRolfRepo extends JpaRepository<VerticesAndArrowsRolf, Long> {
+    List<VerticesAndArrowsRolf> findAllByDiagramIdOrderByStartPointAsc(Long diagramId);
 }
